refactor: Edit some naming problems
